Crew points, missing error messages, form submit save/error feedback, tests

+++ database/migrations/2020_11_23_082915_create_crews_table.php
@@ -15,7 +15,7 @@ class CreateCrewsTable extends Migration
     {
         Schema::create('crews', function (Blueprint $table) {
             $table->id();
-            $table->integer('points');
+            $table->decimal('points');
             $table->integer('person_id');
             $table->integer('title_id');
             $table->integer('media_id');
